Uzair - Commit 4
Mostly Changed AbstOrganism and MyOrganism. Created/changed variables to complement OG brainstorming. Sorted Variables out and added a custom variables section for modding. Also added pseudocode for the MyOrganism functions, and temporary move code which is just WASD for now. 

Laura will review this code and after agreement with rest of group this branch will be pulled to master.